Made some changes

+++ src/js/history.js
@@ -1,10 +1,14 @@
 function historyItemClick(evt, tabId) {
     evt.stopPropagation()
-    getHistory(tabId, (found, data) => {
-        if (found) {
-            createNewReqTab(evt, tabId, data)            
-        }
-    })
+    if (checkTeamIsPersonal()) {
+        getHistory(tabId, (found, data) => {
+            if (found) {
+                createNewReqTab(evt, tabId, data)
+            }
+        })
+    } else {
+        // network
+    }
 }
 
 function toggleHistoryTabs(evt, what) {
@@ -25,37 +29,42 @@ function toggleHistoryTabs(evt, what) {
         .classList.add("tab-content-active")
 }
 
-
 function refreshHistoryTab() {
-    document.querySelector(".history.historyUl")
-        .innerHTML = "<div>Loading history data...</div>"
+    if (checkTeamIsPersonal()) {
+        document.querySelector(".history.historyUl")
+            .innerHTML = "<div>Loading history data...</div>"
     
-    // get all histories belonging to the current team
-    getAllHistoryByTeamId(currentTeam.id, (done, res) => {
-        //log("Loading history:",done, res)
-        if(done) {
-            var htmlStr = ""
-            res.forEach(r => {
-                var h = `
+        // get all histories belonging to the current team
+        getAllHistoryByTeamId(currentTeam.id, (done, res) => {
+            //log("Loading history:",done, res)
+            if (done) {
+                var htmlStr = ""
+                res.forEach(r => {
+                    var h = `
                     <li class="historyLi" onclick="return historyItemClick(event, '${r.tabId}')"><span class="historyItem historyItemMethodType color-orangered bd-rad-3 pd-2">${r.methodType}</span><span class="historyItemName">${r.name}</span></li>
                 `
-                htmlStr += h
-            })
-            document.querySelector(".history.historyUl")
-                .innerHTML = htmlStr
-        }
-    })
+                    htmlStr += h
+                })
+                document.querySelector(".history.historyUl")
+                    .innerHTML = htmlStr
+            }
+        })
+    }
 }
 
 function clearAllHistory() {
-    if(confirm("You sure you want to clear all your history?")) {
-        deleteAllHistory((done) => {
-            if(done) {
-                //log("done")
-            } else {
-                //log("Not done")
-            }
-            refreshHistoryTab()
-        })
+    if (confirm("You sure you want to clear all your history?")) {
+        if (checkTeamIsPersonal()) {
+            deleteAllHistory((done) => {
+                if (done) {
+                    //log("done")
+                } else {
+                    //log("Not done")
+                }
+                refreshHistoryTab()
+            })            
+        } else {
+            // network
+        }
     }
 }
